Modified LoginView
- Modified TextFields
- Add gestureRecognizer to ProfileView
